Alterando e criando o service e o controller da classe transção

+++ src/main/java/com/bancopan/cloud9/bluebank/models/ContaPoupancaModel.java
@@ -1,27 +1,28 @@
 package com.bancopan.cloud9.bluebank.models;
 
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
-import javax.persistence.*;
-
-@Entity(name="tb_conta_poupanca")
+@Entity(name = "tb_conta_poupanca")
 public class ContaPoupancaModel extends ContaModel implements Serializable {
 
-	private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long conta_poupanca;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long conta_poupanca;
 
-	public ContaPoupancaModel() {
-	}
+    public ContaPoupancaModel() {
+    }
 
-	public ContaPoupancaModel(Date dataDeAbertura, BigDecimal saldoContaCorrente) {
-		super(dataDeAbertura, saldoContaCorrente);
-	}
+    public ContaPoupancaModel(Date dataDeAbertura, Double saldoContaCorrente) {
+        super(dataDeAbertura, saldoContaCorrente);
+    }
 
 }
